requirements.txt
keystone-engine
capstone
click
hjson
future
configparser